Reduced testing case further
Left only a single program that works on the 3 B+ just fine, but stalls on the Zero IF the buffers fed to the program are camera frames.
The programs simply uses the TMU to read the buffer, but then discards the values.
Options to the main program include -t to set the amount of programs the buffer is split among, and -m to toggle the way the programs are distributed among the buffer (continuous or interleaved).
And -e, which will replace the camera buffer with a normal VCSM buffer.

Observations:
sudo ./main -c qpu_tmu_read.asm -e		And any combination of other parameters work fine on the Zero
sudo ./main -c qpu_tmu_read.asm			And any combination without -e stall immediately (41/45 instructions in) or after a few frames

Note that -e does NOT stop the camera from running and supplying frames, it just replaces those with emulated buffers before feeding them to the program.